Breaking circular imports: Move command definitions to yubikey_defs
This breaks the yubikey_config <-> yubikey_frame circular import.
