+++ yubico/yubikey_defs.py
@@ -31,3 +31,21 @@ SHA1_DIGEST_SIZE	= 20   # Size of SHA1 digest = 160 bits
 OTP_CHALRESP_SIZE	= 16   # Number of bytes returned for an Yubico-OTP challenge (not from ykdef.h)
 
 UID_SIZE		= 6    # Size of secret ID field
+
+SLOT_CONFIG			= 0x01	# First (default / V1) configuration
+SLOT_CONFIG2			= 0x03	# Second (V2) configuration
+SLOT_UPDATE1			= 0x04	# Update slot 1
+SLOT_UPDATE2			= 0x05	# Update slot 2
+SLOT_SWAP			= 0x06	# Swap slot 1 and 2
+
+def command2str(num):
+    """ Turn command number into name """
+    known = {0x01: "SLOT_CONFIG",
+             0x03: "SLOT_CONFIG2",
+             0x04: "SLOT_UPDATE1",
+             0x05: "SLOT_UPDATE2",
+             0x06: "SLOT_SWAP",
+             }
+    if num in known:
+        return known[num]
+    return "0x%02x" % (num)
